+++ jannovar-vardbs/src/main/java/de/charite/compbio/jannovar/vardbs/dbsnp/DBSNPVariantProperty.java
@@ -1,0 +1,12 @@
+package de.charite.compbio.jannovar.vardbs.dbsnp;
+
+/**
+ * Representation of the dbSNP variant property entry.
+ * 
+ * @author Manuel Holtgrewe <manuel.holtgrewe@bihealth.de>
+ */
+public class DBSNPVariantProperty {
+	
+	// TODO: write me!
+
+}